Mirror package structure in Halyard core
