+++ halyard-config/src/main/java/com/netflix/spinnaker/halyard/config/model/v1/node/Validatable.java
@@ -16,7 +16,7 @@
 
 package com.netflix.spinnaker.halyard.config.model.v1.node;
 
-import com.netflix.spinnaker.halyard.config.model.v1.problem.ConfigProblemSetBuilder;
+import com.netflix.spinnaker.halyard.config.problem.v1.ConfigProblemSetBuilder;
 
 /**
  * Anything implementing this interface is validatable, meaning after we validate it, we get a non-null list of HalconfigProblems
