+++ halyard-config/src/main/java/com/netflix/spinnaker/halyard/config/model/v1/node/Validator.java
@@ -16,7 +16,7 @@
 
 package com.netflix.spinnaker.halyard.config.model.v1.node;
 
-import com.netflix.spinnaker.halyard.config.model.v1.problem.ConfigProblemSetBuilder;
+import com.netflix.spinnaker.halyard.config.problem.v1.ConfigProblemSetBuilder;
 
 public abstract class Validator<T extends Node> {
   abstract public void validate(ConfigProblemSetBuilder p, T n);
