+++ halyard-config/src/main/java/com/netflix/spinnaker/halyard/config/validate/v1/providers/AccountValidator.java
@@ -18,7 +18,7 @@ package com.netflix.spinnaker.halyard.config.validate.v1.providers;
 
 import com.netflix.spinnaker.halyard.config.model.v1.node.Account;
 import com.netflix.spinnaker.halyard.config.model.v1.node.Validator;
-import com.netflix.spinnaker.halyard.config.model.v1.problem.ConfigProblemSetBuilder;
+import com.netflix.spinnaker.halyard.config.problem.v1.ConfigProblemSetBuilder;
 import com.netflix.spinnaker.halyard.core.problem.v1.Problem.Severity;
 import org.springframework.stereotype.Component;
